# Add JIT to flash attention