@lucidrains 2022-07-23
@VHellendoorn oh hey Vincent! thanks for the PR! i don't think this version of flash attention can ever be as performant as Tri's CUDA implementation, but it is good to jit everything :smile: