@GallagherCommaJack 2022-08-24
these results look somewhat suspicious - the standard attention compute time doesn't change with batch size????
really suggests that something weird and overhead bound is happening, or that the benchmark is being measured improperly.