@GallagherCommaJack 2022-08-24
what hardware did you benchmark this on? GPU? TPU?