@VHellendoorn 2022-08-24
Hm, I wasn't so much focused on the standard attention results as on the comparison between Rabe & Flash. From a quick check, the plain one is indeed supposed to be slower, but `@jit` might have played a role by caching results. When I increase the repeats from 10 to 100, the timing results scale linearly:
```
BATCH SIZE: 1
Standard attention: 3.5s
BATCH SIZE: 4
Standard attention: 14.5s
BATCH SIZE: 16
Standard attention: 59.6s
BATCH SIZE: 64
Standard attention: OOM
```
This is a rather disproportionate increase in time (>100x) for a 10x increase in steps, which could be due to JIT, though it might also be a GC effect. It would put Rabe & Flash firmly ahead starting around batch size 4.

In any case, this isn't a refined benchmark by any margin, just a `timeit` loop around an attention call. The goal was to check if Flash attention could be sped up and the finding was that adding a `@jit` call would do so. I'd be happy to see code that more accurately compares Flash with standard attention.